Use int instead of size_t for Vec length/index API
Vec's len/cap/capacityHint members and its index/count/size parameters are
now int, matching the codebase's int-length convention, and the internal
redundant int<->size_t casts are gone. The allocation-size arithmetic
(kElSize, EnsureCapSlow byte math) stays size_t to avoid overflow, and
size_t size() is kept as the STL-compat accessor to avoid churning ~292
call sites. Clean up a few now-redundant casts at call sites.

prompt: change use of size_t in Vec.h to int, update callers and clean up redundant int <=> size_t casts

+++ src/FilterHighlightDraw.cpp
@@ -29,7 +29,7 @@ void DrawMaybeHighlightedText(HDC hdc, RECT rc, Str text, const StrVec& filterWo
 
     // find all match ranges in text
     int textLen = text.len;
-    u8* hl = highlighted.EnsureCap((size_t)textLen);
+    u8* hl = highlighted.EnsureCap(textLen);
     memset(hl, 0, textLen);
     for (int w = 0; w < nWords; w++) {
         Str word = filterWords.At(w);
